% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapminder.R
\docType{data}
\name{gapminder}
\alias{gapminder}
\title{gapminder tibble}
\format{object of class tibble inherited from class table and dataframe}
\source{
gapminder package
}
\usage{
gapminder
}
\description{
gapminder tibble
}
\keyword{datasets}
